refactor(snapshot): establish presentation ownership boundary
