+++ packages/contracts/src/facades/snapshot-presentation.ts
@@ -1,0 +1,6 @@
+export {
+  createSnapshotPresentationNode,
+  foldSnapshotRect,
+  serializeRegularSnapshotPresentationNode,
+} from '../snapshot-presentation.ts';
+export type { SnapshotPresentationNode } from '../snapshot-presentation.ts';
