+++ src/snapshot/snapshot-presentation/ios/action-shelf.test.ts
