+++ src/snapshot/snapshot-presentation/ios/action-shelf.ts
