+++ src/snapshot/snapshot-presentation/ios/mapping.test.ts
