+++ src/snapshot/snapshot-presentation/ios/noise.test.ts
