+++ src/snapshot/snapshot-presentation/ios/publication-membership.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 import type { RawSnapshotNode } from '@agent-device/kernel/snapshot';
-import { buildSnapshotState } from '../../snapshot-state.ts';
+import { buildSnapshotState } from '../../../daemon/snapshot-state.ts';
 
 // End-to-end publication-membership contract for the acquire/present design (#1797, external
 // review pass 4 finding 1), exercised through the production interface. Runner presentation owns
