+++ src/snapshot/snapshot-presentation/ios/rows.fixtures.ts
