+++ src/snapshot/snapshot-presentation/ios/rows.ts
