+++ src/snapshot/snapshot-presentation/ios/scroll.ts
