+++ src/snapshot/snapshot-presentation/ios/transitions.fixtures.ts
